requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0+cu126
torchvision==0.21.0+cu126
torchaudio==2.6.0+cu126
fastapi
uvicorn
streamlit
pymupdf
python-docx
beautifulsoup4
sentence-transformers
faiss-cpu
requests==2.32.5
urllib3==2.2.3
charset-normalizer==3.3.2
chardet==5.2.0
pytest
langgraph
langchain-core
pydantic
python-multipart
langchain-text-splitters
qdrant-client
aiohttp
tiktoken
tenacity
python-dotenv
playwright
pillow
numpy
transformers
coqui-tts
easyocr
celery
redis
accelerate
